Shruthi | #1131 | Adding openelis-atomdfeed-client-beanshell to download script.

+++ scripts/download-build.sh
@@ -1,4 +1,10 @@
 # $1 should be /packages/build folder
+if [ $# -lt 1 ]
+then
+        echo "Usage : $0 download_folder(/packages/build)"
+        exit
+fi
+
 cd $1
 rm -rf *
 wget https://ci-bahmni.thoughtworks.com/go/files/Bahmni_MRS_Pipeline/Latest/BuildDistroStage/Latest/BahmniDistro/openmrs-distro-bahmni-artifacts/distro-1.0-SNAPSHOT-distro.zip
@@ -7,6 +13,7 @@ wget https://ci-bahmni.thoughtworks.com/go/files/Bahmni_MRS_Pipeline/Latest/Buil
 wget https://ci-bahmni.thoughtworks.com/go/files/Bahmni_MRS_Pipeline/Latest/BuildStage/Latest/FunctionalTests/deployables/bahmni-apps.zip
 wget https://ci-bahmni.thoughtworks.com/go/files/Bahmni_MRS_Pipeline/Latest/BuildStage/Latest/FunctionalTests/deployables/jss-old-data-0.2-SNAPSHOT-jar-with-dependencies.jar
 wget https://ci-bahmni.thoughtworks.com/go/files/Bahmni_MRS_Pipeline/Latest/BuildStage/Latest/FunctionalTests/deployables/openmrs-data-jars.zip
+wget https://ci-bahmni.thoughtworks.com/go/files/Bahmni_MRS_Pipeline/Latest/BuildStage/Latest/BahmniCore/elisatomfeedclient-beanshell.zip
 
 OPENElIS_BUILD="Latest"
 wget https://ci-bahmni.thoughtworks.com/go/files/OpenElis/$OPENElIS_BUILD/Build/Latest/build/openelis.war
